Revert "Use logService for console messages (#214)"
This reverts commit 0fed528b6f1fe640acd9c321aa488b425cdf238e.
