+++ src/abstractions/log.service.ts
@@ -6,6 +6,4 @@ export abstract class LogService {
     warning: (message: string) => void;
     error: (message: string) => void;
     write: (level: LogLevelType, message: string) => void;
-    time: (label: string) => void;
-    timeEnd: (label: string) => bigint;
 }
