+++ src/cli/services/consoleLog.service.ts
@@ -1,10 +1,27 @@
 import { LogLevelType } from '../../enums/logLevelType';
 
-import { ConsoleLogService as BaseConsoleLogService } from '../../services/consoleLog.service';
+import { LogService as LogServiceAbstraction } from '../../abstractions/log.service';
 
-export class ConsoleLogService extends BaseConsoleLogService {
-    constructor(isDev: boolean, filter: (level: LogLevelType) => boolean = null) {
-        super(isDev, filter);
+export class ConsoleLogService implements LogServiceAbstraction {
+    constructor(private isDev: boolean, private filter: (level: LogLevelType) => boolean = null) { }
+
+    debug(message: string) {
+        if (!this.isDev) {
+            return;
+        }
+        this.write(LogLevelType.Debug, message);
+    }
+
+    info(message: string) {
+        this.write(LogLevelType.Info, message);
+    }
+
+    warning(message: string) {
+        this.write(LogLevelType.Warning, message);
+    }
+
+    error(message: string) {
+        this.write(LogLevelType.Error, message);
     }
 
     write(level: LogLevelType, message: string) {
@@ -12,12 +29,25 @@ export class ConsoleLogService extends BaseConsoleLogService {
             return;
         }
 
-        if (process.env.BW_RESPONSE) {
-            // tslint:disable-next-line
-            console.error(message);
-            return;
+        switch (level) {
+            case LogLevelType.Debug:
+                // tslint:disable-next-line
+                console.log(message);
+                break;
+            case LogLevelType.Info:
+                // tslint:disable-next-line
+                console.log(message);
+                break;
+            case LogLevelType.Warning:
+                // tslint:disable-next-line
+                console.warn(message);
+                break;
+            case LogLevelType.Error:
+                // tslint:disable-next-line
+                console.error(message);
+                break;
+            default:
+                break;
         }
-
-        super.write(level, message);
     }
 }
