+++ src/electron/services/electronLog.service.ts
@@ -8,8 +8,6 @@ import { LogLevelType } from '../../enums/logLevelType';
 import { LogService as LogServiceAbstraction } from '../../abstractions/log.service';
 
 export class ElectronLogService implements LogServiceAbstraction {
-    private timersMap: Map<string, bigint> = new Map();
-
     constructor(private filter: (level: LogLevelType) => boolean = null, logDir: string = null) {
         if (log.transports == null) {
             return;
@@ -62,18 +60,5 @@ export class ElectronLogService implements LogServiceAbstraction {
             default:
                 break;
         }
-    }
-
-    time(label: string = 'default') {
-        if (!this.timersMap.has(label)) {
-            this.timersMap.set(label, process.hrtime.bigint());
-        }
-    }
-
-    timeEnd(label: string = 'default'): bigint {
-        const elapsed = (process.hrtime.bigint() - this.timersMap.get(label)) / BigInt(1000000);
-        this.timersMap.delete(label);
-        this.write(LogLevelType.Info, `${label}: ${elapsed}ms`);
-        return elapsed;
     }
 }
