+++ src/importers/baseImporter.ts
@@ -13,16 +13,12 @@ import { FolderView } from '../models/view/folderView';
 import { LoginView } from '../models/view/loginView';
 import { SecureNoteView } from '../models/view/secureNoteView';
 
-import { LogService } from '../abstractions/log.service';
 import { CipherType } from '../enums/cipherType';
 import { FieldType } from '../enums/fieldType';
 import { SecureNoteType } from '../enums/secureNoteType';
-import { ConsoleLogService } from '../services/consoleLog.service';
 
 export abstract class BaseImporter {
     organization = false;
-
-    protected logService: LogService = new ConsoleLogService(false);
 
     protected newLineRegex = /(?:\r\n|\r|\n)/;
 
@@ -88,7 +84,7 @@ export abstract class BaseImporter {
             result.errors.forEach((e) => {
                 if (e.row != null) {
                     // tslint:disable-next-line
-                    this.logService.warning('Error parsing row ' + e.row + ': ' + e.message);
+                    console.warn('Error parsing row ' + e.row + ': ' + e.message);
                 }
             });
         }
