+++ src/services/search.service.ts
@@ -5,20 +5,15 @@ import { CipherView } from '../models/view/cipherView';
 import { CipherService } from '../abstractions/cipher.service';
 import { SearchService as SearchServiceAbstraction } from '../abstractions/search.service';
 
-import { LogService } from '../abstractions/log.service';
 import { CipherType } from '../enums/cipherType';
 import { FieldType } from '../enums/fieldType';
 import { UriMatchType } from '../enums/uriMatchType';
-import { ConsoleLogService } from '../services/consoleLog.service';
 
 export class SearchService implements SearchServiceAbstraction {
     private indexing = false;
     private index: lunr.Index = null;
 
-    constructor(private cipherService: CipherService, private logService?: LogService) {
-        if (!logService) {
-            this.logService = new ConsoleLogService(false);
-        }
+    constructor(private cipherService: CipherService) {
     }
 
     clearIndex(): void {
@@ -35,8 +30,8 @@ export class SearchService implements SearchServiceAbstraction {
         if (this.indexing) {
             return;
         }
-
-        this.logService.time('search indexing');
+        // tslint:disable-next-line
+        console.time('search indexing');
         this.indexing = true;
         this.index = null;
         const builder = new lunr.Builder();
@@ -67,8 +62,8 @@ export class SearchService implements SearchServiceAbstraction {
         ciphers.forEach((c) => builder.add(c));
         this.index = builder.build();
         this.indexing = false;
-
-        this.logService.timeEnd('search indexing');
+        // tslint:disable-next-line
+        console.timeEnd('search indexing');
     }
 
     async searchCiphers(query: string,
